requirements.txt
Django==4.2.18
djangorestframework==3.15.2
markdown==3.5
django-filter==24.3
djangorestframework-simplejwt[crypto]==5.4.0
pyyaml==6.0.2
uritemplate==4.1.1
drf-spectacular==0.28.0
pytest-django==4.9.0
pytest-order==1.3.0
factory-boy==3.3.1